Fix Crespo issues with modern Pandas versions
The pd.date_range() argument "closed" was deprecated in Pandas 1.4 with "inclusive" as the preferred replacement. See https://github.com/pandas-dev/pandas/issues/40245.

The expanding() argument "center" was deprecated in Pandas 1.1 as "non-sensical" with no replacement. See https://github.com/pandas-dev/pandas/issues/20647. As discussed in that issue, I believe the use of center=True is a bug in the Crespo implementation, and removing it is the right thing to do.

Both of these deprecated arguments were removed in Pandas 2.0, so this change is needed to make Crespo work with recent Pandas versions.

To ensure no breakage because of this change, I'm adding a pandas>=1.4.0 dependency version constraint to ensure the presence of the "inclusive" argument to pd.date_range().